+++ src/lib.rs
@@ -20,7 +20,7 @@ mod weight_functions;
 pub use convolver::{Convolver, ConvolverFFT};
 pub use functional::{HelmholtzEnergyFunctional, DFT};
 pub use functional_contribution::{FunctionalContribution, FunctionalContributionDual};
-pub use geometry::{Axis, AxisGeometry, Grid};
+pub use geometry::{Axis, Geometry, Grid};
 pub use profile::{DFTProfile, DFTSpecification, DFTSpecifications};
 pub use solver::DFTSolver;
 pub use weight_functions::{WeightFunction, WeightFunctionInfo, WeightFunctionShape};