+++ src/python/adsorption/mod.rs
@@ -1,7 +1,7 @@
 mod external_potential;
 mod pore;
 
-pub use external_potential::{PyExternalPotential, PyGeometry};
+pub use external_potential::PyExternalPotential;
 
 #[macro_export]
 macro_rules! impl_adsorption {