+++ CHANGELOG.md
@@ -6,13 +6,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [0.2.0] - 2022-03-09
+## [0.2.0] - 2022-03-??
+### Changed
+- Renamed `AxisGeometry` to `Geometry`.
+- Removed `PyGeometry` and `PyFMTVersion` in favor of a simpler implementation using `PyO3`'s new `#[pyclass]` for fieldless enums feature.
+- `DFTSolver` now uses `Verbosity` instead of a `bool` to control its output.
+
 ### Packaging
 - Updated `pyo3` and `numpy` dependencies to 0.16.
 - Updated `quantity` dependency to 0.5.
 - Updated `num-dual` dependency to 0.5.
 - Updated `feos-core` dependency to 0.2.
 - Updated `ang` dependency to 0.6.
+- Removed `log` dependency.
 
 ## [0.1.3] - 2022-02-17
 ### Fixed